requirements.txt
django==2.1
django-debug-toolbar==1.8
gunicorn==19.4.5
psycopg2==2.7.3.1
whitenoise==3.3.1
django-cron==0.5.1
pillow==5.4.1
django_crispy_forms==1.8.1
